Move match-based builtins code to bench.rs
