+++ benches/bench.rs
@@ -6,6 +6,11 @@ extern crate meval;
 use test::Bencher;
 use meval::Expr;
 use meval::Context;
+use meval::FuncEvalError;
+use meval::min_array;
+use meval::max_array;
+use meval::ContextProvider;
+use std::f64::consts;
 
 const EXPR: &'static str = "abs(sin(x + 1) * (x^2 + x + 1))";
 
@@ -17,9 +22,9 @@ fn parsing(b: &mut Bencher) {
 }
 
 #[bench]
-fn evaluation(b: &mut Bencher) {
+fn evaluation_matchcontext(b: &mut Bencher) {
     let expr = Expr::from_str(EXPR).unwrap();
-    let func = expr.bind("x").unwrap();
+    let func = expr.bind_with_context(MatchBuiltins, "x").unwrap();
     b.iter(|| {
         func(1.);
     });
@@ -32,4 +37,78 @@ fn evaluation_hashcontext(b: &mut Bencher) {
     b.iter(|| {
         func(1.);
     });
+}
+
+macro_rules! one_arg {
+    ($args:expr, $func:ident) => {
+        if $args.len() == 1 {
+            Ok($args[0].$func())
+        } else {
+            Err(FuncEvalError::NumberArgs(1))
+        }
+    }
+}
+
+macro_rules! two_args {
+    ($args:expr, $func:ident) => {
+        if $args.len() == 2 {
+            Ok($args[0].$func($args[1]))
+        } else {
+            Err(FuncEvalError::NumberArgs(2))
+        }
+    }
+}
+
+macro_rules! one_or_more_arg {
+    ($args:expr, $func:ident) => {
+        if $args.len() >= 1 {
+            Ok($func($args))
+        } else {
+            Err(FuncEvalError::TooFewArguments)
+        }
+    }
+}
+
+/// Built-in functions and constants.
+///
+/// See the library documentation for the list of built-ins.
+#[doc(hidden)]
+pub struct MatchBuiltins;
+
+impl ContextProvider for MatchBuiltins {
+    fn get_var(&self, name: &str) -> Option<f64> {
+        match name {
+            "pi" => Some(consts::PI),
+            "e" => Some(consts::E),
+            _ => None,
+        }
+    }
+    fn eval_func(&self, name: &str, args: &[f64]) -> Result<f64, FuncEvalError> {
+        match name {
+            "sqrt" => one_arg!(args, sqrt),
+            "exp" => one_arg!(args, exp),
+            "ln" => one_arg!(args, ln),
+            "abs" => one_arg!(args, abs),
+            "sin" => one_arg!(args, sin),
+            "cos" => one_arg!(args, cos),
+            "tan" => one_arg!(args, tan),
+            "asin" => one_arg!(args, asin),
+            "acos" => one_arg!(args, acos),
+            "atan" => one_arg!(args, atan),
+            "sinh" => one_arg!(args, sinh),
+            "cosh" => one_arg!(args, cosh),
+            "tanh" => one_arg!(args, tanh),
+            "asinh" => one_arg!(args, asinh),
+            "acosh" => one_arg!(args, acosh),
+            "atanh" => one_arg!(args, atanh),
+            "floor" => one_arg!(args, floor),
+            "ceil" => one_arg!(args, ceil),
+            "round" => one_arg!(args, round),
+            "signum" => one_arg!(args, signum),
+            "atan2" => two_args!(args, atan2),
+            "max" => one_or_more_arg!(args, max_array),
+            "min" => one_or_more_arg!(args, min_array),
+            _ => Err(FuncEvalError::UnknownFunction),
+        }
+    }
 }
