+++ src/expr.rs
@@ -368,87 +368,16 @@ impl std::error::Error for FuncEvalError {
     }
 }
 
-/// Built-in functions and constants.
-///
-/// See the library documentation for the list of built-ins.
 #[doc(hidden)]
-pub struct Builtins;
-
-macro_rules! one_arg {
-    ($args:expr, $func:ident) => {
-        if $args.len() == 1 {
-            Ok($args[0].$func())
-        } else {
-            Err(FuncEvalError::NumberArgs(1))
-        }
-    }
-}
-
-macro_rules! two_args {
-    ($args:expr, $func:ident) => {
-        if $args.len() == 2 {
-            Ok($args[0].$func($args[1]))
-        } else {
-            Err(FuncEvalError::NumberArgs(2))
-        }
-    }
-}
-
-macro_rules! one_or_more_arg {
-    ($args:expr, $func:ident) => {
-        if $args.len() >= 1 {
-            Ok($func($args))
-        } else {
-            Err(FuncEvalError::TooFewArguments)
-        }
-    }
-}
-
-fn max_array(xs: &[f64]) -> f64 {
+pub fn max_array(xs: &[f64]) -> f64 {
     xs.iter().fold(::std::f64::NEG_INFINITY, |m, &x| m.max(x))
 }
 
-fn min_array(xs: &[f64]) -> f64 {
+#[doc(hidden)]
+pub fn min_array(xs: &[f64]) -> f64 {
     xs.iter().fold(::std::f64::INFINITY, |m, &x| m.min(x))
 }
 
-impl ContextProvider for Builtins {
-    fn get_var(&self, name: &str) -> Option<f64> {
-        match name {
-            "pi" => Some(consts::PI),
-            "e" => Some(consts::E),
-            _ => None,
-        }
-    }
-    fn eval_func(&self, name: &str, args: &[f64]) -> Result<f64, FuncEvalError> {
-        match name {
-            "sqrt" => one_arg!(args, sqrt),
-            "exp" => one_arg!(args, exp),
-            "ln" => one_arg!(args, ln),
-            "abs" => one_arg!(args, abs),
-            "sin" => one_arg!(args, sin),
-            "cos" => one_arg!(args, cos),
-            "tan" => one_arg!(args, tan),
-            "asin" => one_arg!(args, asin),
-            "acos" => one_arg!(args, acos),
-            "atan" => one_arg!(args, atan),
-            "sinh" => one_arg!(args, sinh),
-            "cosh" => one_arg!(args, cosh),
-            "tanh" => one_arg!(args, tanh),
-            "asinh" => one_arg!(args, asinh),
-            "acosh" => one_arg!(args, acosh),
-            "atanh" => one_arg!(args, atanh),
-            "floor" => one_arg!(args, floor),
-            "ceil" => one_arg!(args, ceil),
-            "round" => one_arg!(args, round),
-            "signum" => one_arg!(args, signum),
-            "atan2" => two_args!(args, atan2),
-            "max" => one_or_more_arg!(args, max_array),
-            "min" => one_or_more_arg!(args, min_array),
-            _ => Err(FuncEvalError::UnknownFunction),
-        }
-    }
-}
 
 /// Returns the built-in constants and functions in a form that can be used as a `ContextProvider`.
 #[doc(hidden)]
@@ -722,8 +651,8 @@ mod tests {
                                          Context::new().func3("phi", |x, y, z| x + y * z)),
                    Ok(2. + 3. * 4.));
         assert_eq!(eval_str_with_context("phi(2., 3.)",
-                                         Context::new().funcn("phi",
-                                                              |xs: &[f64]| xs[0] + xs[1], 2)),
+                                         Context::new()
+                                             .funcn("phi", |xs: &[f64]| xs[0] + xs[1], 2)),
                    Ok(2. + 3.));
         let mut m = HashMap::new();
         m.insert("x", 2.);
